requirements.txt
numpy==1.26.1
pyserial==3.5
matplotlib==3.8.0
tqdm==4.66.1
scikit-learn~=1.3.0
scipy>=1.11.3
keras==2.10.0